% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_df.R
\name{aasset_to_df}
\alias{aasset_to_df}
\title{AAStringset class to dataframe}
\usage{
aasset_to_df(aas)
}
\arguments{
\item{aas}{AAStringset class object.}
}
\value{
Data frame.
}
\description{
AAStringset class to dataframe
}
