% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_deepredeff.R
\name{autoplot.heatmap_deepredeff}
\alias{autoplot.heatmap_deepredeff}
\title{Plot the heatmap}
\usage{
\method{autoplot}{heatmap_deepredeff}(object, n = 50, ...)
}
\arguments{
\item{object}{Results from deepredeff::get_sum_activation().}

\item{n}{Number of harmonics.}

\item{...}{Additional arguments ignored.}
}
\value{
Plot of raw heatmap data and its smooth version.
}
\description{
Plot the heatmap
}
