% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeffpred.R
\docType{package}
\name{deeffpred}
\alias{deeffpred}
\title{\code{deeffpred} package}
\description{
Effector protein predictor using Deep Learning models.
}
\details{
See the README on
\href{https://github.com/ruthkr/deeffpred/}{GitHub}
}
