% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepredeff.R
\docType{package}
\name{deepredeff}
\alias{deepredeff}
\title{\code{deepredeff} package}
\description{
Effector protein predictor using Deep Learning models.
}
\details{
See the README on
\href{https://github.com/ruthkr/deepredeff/}{GitHub}
}
