% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_methods.R
\name{encode_integer}
\alias{encode_integer}
\title{Integer encoder}
\usage{
encode_integer(sequence, max_length = 4034)
}
\arguments{
\item{sequence}{Sequence.}

\item{max_length}{Maximum sequence length used on the training data.}
}
\value{
Integer encoded sequence.
}
\description{
Integer encoder
}
