% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_one_hot.R
\name{encode_one_hot}
\alias{encode_one_hot}
\title{TOne-hot Encoder}
\usage{
encode_one_hot(sequence, max_length = 4034)
}
\arguments{
\item{sequence}{Sequence}

\item{max_length}{Maximum sequence length of the used training data}
}
\value{
Encoded sequence
}
\description{
TOne-hot Encoder
}
