% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_methods.R
\name{ensemble_voting}
\alias{ensemble_voting}
\title{Ensemble voting}
\usage{
ensemble_voting(pred_list, weights = NULL)
}
\arguments{
\item{pred_list}{List of deep learning models.}

\item{weights}{Dummy weights variable.}
}
\value{
Binary prediction.
}
\description{
Ensemble voting
}
