% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_convnet_helper.R
\name{get_class_activation_each_seq}
\alias{get_class_activation_each_seq}
\title{Get class activation value for each sequence}
\usage{
get_class_activation_each_seq(data, loaded_model, layer)
}
\arguments{
\item{data}{An encoded sequence as input of prediction.}

\item{loaded_model}{Model to use to predict.}

\item{layer}{Convnet layer to use.}
}
\value{
heatmap
}
\description{
This function is for getting the class activation value, which indicates how important each location with the respect to the class under consideration.
}
