% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_convnet.R
\name{get_class_activation_value}
\alias{get_class_activation_value}
\title{Get class activation value}
\usage{
get_class_activation_value(data, model, layer)
}
\arguments{
\item{data}{Encoded sequence data as input of prediction.}

\item{model}{Model to use to predict.}

\item{layer}{Convnet layer to use.}
}
\value{
heatmap
}
\description{
This function is for getting the class activation value, which indicates how important each location with the respect to the class under consideration.
}
