% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_convnet_helper.R
\name{get_fft_smooth}
\alias{get_fft_smooth}
\title{Get the smooth function using fourier transformation}
\usage{
get_fft_smooth(x = NULL, n = NULL)
}
\arguments{
\item{x}{Input data.}

\item{n}{Number of harmonics.}
}
\value{
Smooth input using FFT.
}
\description{
Get the smooth function using fourier transformation
}
