% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_convnet_helper.R
\name{get_heatmap_value}
\alias{get_heatmap_value}
\title{Visualise convnet}
\usage{
get_heatmap_value(input_data, taxon)
}
\arguments{
\item{input_data}{amino acid protein sequence input}

\item{taxon}{taxon of the data}
}
\value{
smoothed value of convnet visualisation
}
\description{
Visualise convnet
}
