% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_convnet.R
\name{load_model_to_visualise}
\alias{load_model_to_visualise}
\title{Load model that will be visualised}
\usage{
load_model_to_visualise(model = c("all", "bacteria", "fungi", "oomycete"))
}
\arguments{
\item{model}{type model to visualise}
}
\value{
loaded model
}
\description{
Load model that will be visualised
}
