% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_glob}
\alias{package_glob}
\title{Wildcard Expansion on File Paths}
\usage{
package_glob(..., pattern)
}
\arguments{
\item{...}{Path}

\item{pattern}{Pattern}
}
\value{
Glob
}
\description{
Wildcard Expansion on File Paths
}
