% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_deepredeff.R
\name{plot.heatmap_deepredeff}
\alias{plot.heatmap_deepredeff}
\title{Plot the heatmap}
\usage{
\method{plot}{heatmap_deepredeff}(x, n = 50, ...)
}
\arguments{
\item{x}{The heatmap value results.}

\item{n}{Number of harmonics.}

\item{...}{Additional arguments ignored.}
}
\value{
Plot of raw heatmap data and its smooth version.
}
\description{
Plot the heatmap
}
