% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_deepredeff.R
\name{plot.tbl_deepredeff}
\alias{plot.tbl_deepredeff}
\title{Plot the results of prediction}
\usage{
\method{plot}{tbl_deepredeff}(x, ...)
}
\arguments{
\item{x}{tbl_deepredeff object}

\item{...}{additional arguments ignored.}
}
\value{
class distribution plot
}
\description{
Plot the results of prediction
}
