% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_effector.R
\name{predict_effector}
\alias{predict_effector}
\title{Predict Effector}
\usage{
predict_effector(x, pathogen = "all", col_names)
}
\arguments{
\item{x}{FASTA path or dataframe}

\item{pathogen}{Pathogen}

\item{col_names}{Optional column names for dataframe. One of them must be named 'sequence'}
}
\value{
Dataframe with sequences and effector predictions
}
\description{
Predict Effector
}
