% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_single_effector.R
\name{predict_single_effector}
\alias{predict_single_effector}
\title{Predict on a single sequence}
\usage{
predict_single_effector(sequence, model)
}
\arguments{
\item{sequence}{Sequence}

\item{model}{Trained model}
}
\value{
Prediction
}
\description{
Predict on a single sequence
}
