% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_deepredeff.R
\name{summary.tbl_deepredeff}
\alias{summary.tbl_deepredeff}
\title{Get the summary of the prediction results}
\usage{
\method{summary}{tbl_deepredeff}(object, ...)
}
\arguments{
\item{object}{Results of prediction from deepredeff::predict_effector().}

\item{...}{Additional arguments ignored.}
}
\description{
Get the summary of the prediction results
}
