% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_convnet.R
\name{visualise_convnet}
\alias{visualise_convnet}
\alias{visualise_convnet.character}
\alias{visualise_convnet.data.frame}
\alias{visualise_convnet.AAStringSet}
\alias{visualise_convnet.AAString}
\alias{visualise_convnet.default}
\title{Visualise Convnet}
\usage{
visualise_convnet(input, taxon)

\method{visualise_convnet}{character}(input, taxon)

\method{visualise_convnet}{data.frame}(input, taxon)

\method{visualise_convnet}{AAStringSet}(input, taxon)

\method{visualise_convnet}{AAString}(input, taxon)

\method{visualise_convnet}{default}(input, taxon)
}
\arguments{
\item{input}{Input data that contains amino acid sequence(s). It can be in fasta format, strings, AAString, AAStringset, and dataframe.}

\item{taxon}{Taxon group of input data. Available taxons are bacteria, fungi, and oomycete.}
}
\value{
\code{visualise_convnet} returns an object of class "heatmap_deepredeff" or for multiple responses of class c("heatmap_deepredeff", "array").
}
\description{
\code{visualise_convnet} is used to visualise convolutional network filters while predicting effector proteins given amino acid protein sequences.
}
