% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_df.R
\name{aas_to_df}
\alias{aas_to_df}
\title{Convert AAString class to dataframe}
\usage{
aas_to_df(aas)
}
\arguments{
\item{aas}{AAString class object.}
}
\value{
Returns data in data frame.
}
\description{
Convert AAString class to dataframe
}
\examples{
input_fasta <- system.file("extdata/example/fungi_sample.fasta", package = "deepredeff")
input_aas <- Biostrings::readAAStringSet(input_fasta)
input_df <- aasset_to_df(input_aas)
}
