% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_df.R
\name{aasset_to_df}
\alias{aasset_to_df}
\title{Convert AAStringset class to dataframe}
\usage{
aasset_to_df(aas)
}
\arguments{
\item{aas}{AAStringset class object.}
}
\value{
Returns data in data frame.
}
\description{
Convert AAStringset class to dataframe
}
\examples{
input <- "MSHMTFNTWKAGLWRLAAAAVLSLLPVVARAAVPGITGPTFDLTAQPGRANQPDGASVYSWGYGCNPRTVPGFLPSVNPLAGQ"
input_aas <- Biostrings::AAString(input)
input_df <- aas_to_df(input_aas)
}
