% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_methods.R
\name{encode_integer}
\alias{encode_integer}
\title{Integer encoder}
\usage{
encode_integer(sequence, max_length = 4034)
}
\arguments{
\item{sequence}{Sequence in a string format.}

\item{max_length}{Maximum length of sequence to encode.}
}
\value{
Integer encoded sequence.
}
\description{
\code{encode_integer} integer-encodes sequence in a string format.
}
\examples{
sample_seq <- "MSHMTFNTWKAGLWRLAAAAVLSLLPVVARAAVPGITGPTFDLTAQPGRANQPDGASVYSWGYGCNPRTVPGFLPSVNPLAGQ"
encoded_seq <- encode_integer(sample_seq)
}
