% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_methods.R
\name{encode_one_hot}
\alias{encode_one_hot}
\title{One-hot encoder}
\usage{
encode_one_hot(sequence, max_length = 4034)
}
\arguments{
\item{sequence}{Sequence in a string format.}

\item{max_length}{Maximum length of sequence to encode.}
}
\value{
One-hot encoded sequence.
}
\description{
\code{encode_one_hot} one-hot-encodes sequence in a string format.
}
\examples{
sample_seq <- "MSHMTFNTWKAGLWRLAAAAVLSLLPVVARAAVPGITGPTFDLTAQPGRANQPDGASVYSWGYGCNPRTVPGFLPSVNPLAGQ"
encoded_seq <- encode_one_hot(sample_seq)
}
