% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_methods.R
\name{ensemble_weighted}
\alias{ensemble_weighted}
\title{Weighted ensemble}
\usage{
ensemble_weighted(pred_list, weights)
}
\arguments{
\item{pred_list}{List of deep learning models.}

\item{weights}{Accuracy values from evaluation on the validation dataset.}
}
\value{
Returns the prediction results from weighted ensemble.
}
\description{
Weighted ensemble
}
