% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_to_df.R
\name{fasta_to_df}
\alias{fasta_to_df}
\title{FASTA to dataframe}
\usage{
fasta_to_df(fasta_path)
}
\arguments{
\item{fasta_path}{Path to FASTA file.}
}
\value{
Data frame.
}
\description{
FASTA to dataframe
}
