% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_to_df.R
\name{fasta_to_df}
\alias{fasta_to_df}
\title{Convert fasta format to dataframe}
\usage{
fasta_to_df(fasta_path)
}
\arguments{
\item{fasta_path}{Path of FASTA file.}
}
\value{
Returns data in data frame.
}
\description{
Convert fasta format to dataframe
}
\examples{
input <- system.file("extdata/example/fungi_sample.fasta", package = "deepredeff")
input_df <- fasta_to_df(input)
}
