% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_methods.R
\name{get_ensemble_method}
\alias{get_ensemble_method}
\title{Get ensemble methods}
\usage{
get_ensemble_method(taxon)
}
\arguments{
\item{taxon}{taxon group}
}
\value{
Returns ensemble method and weights.
}
\description{
This function is used to get the ensemble methods used for each taxon group. If weights are needed for a particular ensemble, then the weights will automatically follow.
}
