% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_convnet.R
\name{get_sum_activation}
\alias{get_sum_activation}
\alias{get_sum_activation.character}
\alias{get_sum_activation.data.frame}
\alias{get_sum_activation.AAString}
\alias{get_sum_activation.AAStringSet}
\alias{get_sum_activation.default}
\title{Get sum activation class}
\usage{
get_sum_activation(input, model)

\method{get_sum_activation}{character}(input, model)

\method{get_sum_activation}{data.frame}(input, model)

\method{get_sum_activation}{AAString}(input, model)

\method{get_sum_activation}{AAStringSet}(input, model)

\method{get_sum_activation}{default}(input, model)
}
\arguments{
\item{input}{Input sequence data.}

\item{model}{Model to use.}
}
\value{
Sum of class activation.
}
\description{
This function will calculate sum of all each class activation value in a set of sequence data.
}
