% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Load model in HDF5 format}
\usage{
load_model(model = c("all", "bacteria", "fungi", "oomycete"))
}
\arguments{
\item{model}{Name of model.}
}
\value{
all of the hyperparamaters and parameters of particula model from specific chosen model.
}
\description{
Load model in HDF5 format
}
