% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Load model in HDF5 format}
\usage{
load_model(taxon = c("bacteria", "fungi", "oomycete"))
}
\arguments{
\item{taxon}{Name of taxon.}
}
\value{
all of the hyperparamaters and parameters of particular model from specific chosen taxon.
}
\description{
Load model in HDF5 format
}
