% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Load model}
\usage{
load_model(taxon = c("bacteria", "fungi", "oomycete"))
}
\arguments{
\item{taxon}{Name of taxon.}
}
\value{
Returns all of the hyperparamaters and parameters of particular model from specific taxon chosen.
}
\description{
\code{load_model()} loads model saved in hdf5 format
}
