% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_deepredeff.R
\name{plot.tbl_deepredeff}
\alias{plot.tbl_deepredeff}
\title{Plot the results of prediction}
\usage{
\method{plot}{tbl_deepredeff}(x, ...)
}
\arguments{
\item{x}{tbl_deepredeff object}

\item{...}{additional arguments ignored.}
}
\value{
class distribution plot
}
\description{
Plot the results of prediction
}
\examples{
\donttest{
# FASTA input
input_fasta <- system.file("extdata/example/fungi_sample.fasta", package = "deepredeff")

pred_result <- deepredeff::predict_effector(
  input = input_fasta,
  taxon = "fungi"
)

plot(pred_result)
}

}
