% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_effector}
\alias{predict_effector}
\title{Predict effector}
\usage{
predict_effector(input, model = "all", col_names)
}
\arguments{
\item{input}{FASTA path or dataframe that contains input sequences.}

\item{model}{Selected model.}

\item{col_names}{Optional column names for dataframe. One of them must be named 'sequence'.}
}
\value{
Dataframe with sequences and effector predictions.
}
\description{
Predict effector
}
