% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_effector}
\alias{predict_effector}
\alias{predict_effector.character}
\alias{predict_effector.data.frame}
\alias{predict_effector.AAStringSet}
\alias{predict_effector.AAString}
\alias{predict_effector.default}
\title{Predict effector}
\usage{
predict_effector(input, model)

\method{predict_effector}{character}(input, model)

\method{predict_effector}{data.frame}(input, model)

\method{predict_effector}{AAStringSet}(input, model)

\method{predict_effector}{AAString}(input, model)

\method{predict_effector}{default}(input, model)
}
\arguments{
\item{input}{Input data that contains amino acid sequence(s). It can be in fasta format, strings, AAString, AAStringset, and dataframe.}

\item{model}{Model to use in predicting the input sequence. Available models are bacteria, fungi, oomycete, and all.}
}
\value{
Deepredeff objects that contains sequences and effector predictions.
}
\description{
Predict effector
}
