% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_effector}
\alias{predict_effector}
\alias{predict_effector.character}
\alias{predict_effector.data.frame}
\alias{predict_effector.AAStringset}
\alias{predict_effector.default}
\title{Predict effector}
\usage{
predict_effector(input, model)

\method{predict_effector}{character}(input, model)

\method{predict_effector}{data.frame}(input, model)

\method{predict_effector}{AAStringset}(input, model)

\method{predict_effector}{default}(input, model)
}
\arguments{
\item{input}{FASTA path or dataframe that contains input sequences.}

\item{model}{Selected model.}
}
\value{
Dataframe with sequences and effector predictions.
}
\description{
Predict effector
}
