% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_helper.R
\name{prediction_mapper}
\alias{prediction_mapper}
\title{Prediction mapper helper}
\usage{
prediction_mapper(sequence_list, model_list)
}
\arguments{
\item{sequence_list}{List of sequences input.}

\item{model_list}{List of models.}
}
\value{
Returns list of prediction result of each sequence.
}
\description{
Prediction mapper helper
}
