% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_deepredeff.R
\name{summary.tbl_deepredeff}
\alias{summary.tbl_deepredeff}
\title{Get the summary of the prediction results}
\usage{
\method{summary}{tbl_deepredeff}(object, cutoff = 0.5, ...)
}
\arguments{
\item{object}{Results of prediction from deepredeff::predict_effector().}

\item{cutoff}{Cutoff value to determine the prediction class (1 for effector and 0 for non-effector). The default value is 0.5.}

\item{...}{Additional arguments ignored.}
}
\description{
Get the summary of the prediction results
}
