% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_deepredeff.R
\name{summary.tbl_deepredeff}
\alias{summary.tbl_deepredeff}
\title{Get the summary of the prediction results}
\usage{
\method{summary}{tbl_deepredeff}(object, ...)
}
\arguments{
\item{object}{Results of prediction from deepredeff::predict_effector().}

\item{...}{Additional arguments ignored.}
}
\description{
Get the summary of the prediction results
}
\examples{
\donttest{
# FASTA input
input_fasta <- system.file("extdata/example/fungi_sample.fasta", package = "deepredeff")

pred_result <- deepredeff::predict_effector(
  input = input_fasta,
  taxon = "fungi"
)

summary(pred_result)
}

}
